backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.7

# Resume parsing
pypdf2>=3.0.0
python-docx>=0.8.11
spacy>=3.5.2

# API integrations
requests>=2.28.2
python-dotenv>=1.0.0

# Email sending
sendgrid>=6.10.0

# AI Scoring
google-generativeai>=0.7.0